Extract TypeConverter from mixin.rb
